101 print number

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -6,7 +6,7 @@
 
 void print_number(int n)
 {
-	unsigned int num = n;
+	long num = n;
 
 	if (num < 0)
 	{
